refactor: move unit tests to src/tests.rs
